Move Ontology parser to DG module
A Prove module will be added for parsing of the prove-command output.
